Added more rules

+++ src/evoalgo/tracker/IPointAwarder.java
@@ -4,5 +4,5 @@ public interface IPointAwarder {
 
 	int awardPoints(boolean[] sv,int o);
 
-	double pointsForEffort(boolean[] sv, double move);
+	double pointsForEffort(boolean[] sv, int move);
 }
